+++ example/lib/main.dart
@@ -23,10 +23,20 @@ class _MyAppState extends State<MyApp> {
     super.initState();
   }
 
-  void _resultMessage(String message, {bool error = false}) {
+  void _resultMessage(String message, TransactionStatus status) {
     setState(() {
       _message = message;
-      _color = error ? Colors.red : Colors.green;
+      switch (status) {
+        case TransactionStatus.success:
+          _color = Colors.green;
+          break;
+        case TransactionStatus.userCanceled:
+          _color = Colors.orange;
+          break;
+        case TransactionStatus.fail:
+          _color = Colors.red;
+          break;
+      }
     });
   }
 
@@ -34,9 +44,18 @@ class _MyAppState extends State<MyApp> {
     try {
       ActivationResponse response =
           await _pos.activatePos(apikey: 'API KEY', apiSecret: 'API SECREET');
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _getActivationCode() async {
+    try {
+      GetActivationCodeResponse response = await _pos.getActivationCode();
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
@@ -49,9 +68,9 @@ class _MyAppState extends State<MyApp> {
         showReceipt: false,
         showTransactionResult: false,
       );
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
@@ -63,18 +82,66 @@ class _MyAppState extends State<MyApp> {
         showReceipt: false,
         showTransactionResult: false,
       );
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _fastRefund() async {
+    try {
+      FastRefundResponse response = await _pos.fastRefund(
+        amount: 10.00,
+        showRating: false,
+        showReceipt: false,
+        showTransactionResult: false,
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _transactionDetails() async {
+    try {
+      TransactionResponse response = await _pos.transactionDetails(
+        clientTransactiodId: '1234567890123456789',
+        sourceTerminalId: '12345',
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _reprintTransaction() async {
+    try {
+      ReprintTransactionResponse response = await _pos.reprintTransaction(
+        orderCode: '1234567890123456789',
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _batch() async {
+    try {
+      BatchResponse response = await _pos.batch(
+        command: BatchCommand.open,
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
   void _abort() async {
     try {
       AbortResponse response = await _pos.abort();
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
@@ -84,9 +151,31 @@ class _MyAppState extends State<MyApp> {
         mode: ApplicationMode.attended,
         pin: '1234',
       );
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _setDecimalAmountMode() async {
+    try {
+      SetDecimalAmountModeResponse response = await _pos.setDecimalAmountMode(
+        decimalMode: false,
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
+    }
+  }
+
+  void _resetTerminal() async {
+    try {
+      ResetTerminalResponse response = await _pos.resetTerminal(
+        softReset: true,
+      );
+      _resultMessage(response.message, response.status);
+    } catch (e) {
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
@@ -96,18 +185,18 @@ class _MyAppState extends State<MyApp> {
         businessDescriptionEnabled: true,
         businessDescriptionType: BusinessDescriptionType.storeName,
       );
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
   void _sendLogs() async {
     try {
       SendLogsResponse response = await _pos.sendLogs();
-      _resultMessage(response.message);
+      _resultMessage(response.message, response.status);
     } catch (e) {
-      _resultMessage(e.toString(), error: true);
+      _resultMessage(e.toString(), TransactionStatus.fail);
     }
   }
 
@@ -118,6 +207,7 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('VW Pos Plugin example'),
         ),
+        // ovdje ubaciti fiksni message da bude pri vrhu a dolje LisView scrollable...
         body: ListView(children: <Widget>[
           Padding(
             padding: const EdgeInsets.all(10),
@@ -147,6 +237,17 @@ class _MyAppState extends State<MyApp> {
             height: 70,
             padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
             child: ElevatedButton(
+              child: const Text('Get Activation Code',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () async {
+                _getActivationCode();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
               child: const Text('Sale 10.00', style: TextStyle(fontSize: 18)),
               onPressed: () {
                 _sale();
@@ -157,10 +258,54 @@ class _MyAppState extends State<MyApp> {
             height: 70,
             padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
             child: ElevatedButton(
-              child: const Text('Cancel - refund 10.00',
+              child: const Text('Cancel/refund -10.00',
                   style: TextStyle(fontSize: 18)),
               onPressed: () {
                 _cancel();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
+              child: const Text('Fast refund -10.00',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _fastRefund();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
+              child: const Text('Transaction details',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _transactionDetails();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
+              child: const Text('Reprint transaction',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _reprintTransaction();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
+              child: const Text('Open / close batch',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _batch();
               },
             ),
           ),
@@ -189,6 +334,17 @@ class _MyAppState extends State<MyApp> {
             height: 70,
             padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
             child: ElevatedButton(
+              child: const Text('Set Decimal Amount Mode',
+                  style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _setDecimalAmountMode();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
               child: const Text('Send logs', style: TextStyle(fontSize: 18)),
               onPressed: () {
                 _sendLogs();
@@ -203,6 +359,17 @@ class _MyAppState extends State<MyApp> {
                   style: TextStyle(fontSize: 18)),
               onPressed: () {
                 _setPrintingSettings();
+              },
+            ),
+          ),
+          Container(
+            height: 70,
+            padding: const EdgeInsets.fromLTRB(10, 10, 10, 0),
+            child: ElevatedButton(
+              child:
+                  const Text('Reset Terminal', style: TextStyle(fontSize: 18)),
+              onPressed: () {
+                _resetTerminal();
               },
             ),
           ),