+++ lib/src/types.dart
@@ -5,7 +5,7 @@
 /// Enum for application mode settings
 ///
 /// (https://developer.vivawallet.com/apis-for-point-of-sale/card-terminal-apps/android-app/set-mode/)
-enum ApplicationMode { attended, semiUnattended, fullUnattended }
+enum ApplicationMode { attended, semiUnattended, fullUnattended, ecrControlled }
 
 ///
 /// Merchant’s Business/Trade/Store name (depending on what option is selected in the 'viva.com | Terminal' application settings)
@@ -16,3 +16,8 @@ enum BusinessDescriptionType { businessName, tradeName, storeName }
 /// The status of the transaction
 ///
 enum TransactionStatus { fail, success, userCanceled }
+
+///
+/// Batch request action
+///
+enum BatchCommand { open, close }